almost done

+++ Backend/auth_service/src/main/java/com/ust/auth_service/repository/UserRepository.java
@@ -1,10 +1,11 @@
 package com.ust.auth_service.repository;
 
+import com.ust.auth_service.dto.UserSignupDto;
 import com.ust.auth_service.model.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserModel, Long> {
+public interface UserRepository extends JpaRepository<UserModel, Integer> {
     Optional<UserModel> findByUsername(String username);
 }
